Former-commit-id: 3e344eaf9da164dd2192e8a2fd38a48757cb1b40
Former-commit-id: 163f16d096717f9b96e4a533705a29c770866711

+++ src/core/base64.h
@@ -7,7 +7,7 @@ namespace phi
     class CORE_API base64
     {
     public:
-        static string encode(byte const* buf, uint bufLen);
+        static string encode(byte const* buf, size_t bufLen);
         static vector<byte> decode(string const&);
     };
 }
